backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
algosdk==2.7.0
py-algorand-sdk==2.8.0
pyteal==0.26.0
requests==2.32.3
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
python-dotenv==1.0.1
pytest==8.3.3
